requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
groq==0.4.0
python-dotenv==1.0.0
jinja2==3.1.3
python-multipart==0.0.9
